Code cleanup

+++ Cinesta/WebApp/AppDataHelper.cs
@@ -1,5 +1,4 @@
 ﻿using App.DAL.EF;
-using App.Domain.Cast;
 using Microsoft.EntityFrameworkCore;
 
 namespace WebApp;
@@ -11,10 +10,7 @@ public static class AppDataHelper
         using var serviceScope = app.ApplicationServices.GetRequiredService<IServiceScopeFactory>().CreateScope();
         using var context = serviceScope.ServiceProvider.GetService<AppDbContext>();
 
-        if (context == null)
-        {
-            throw new ApplicationException("Services error. No DB context");
-        }
+        if (context == null) throw new ApplicationException("Services error. No DB context");
 
         //TODO - check  database state
         //can't connect - wrong address
@@ -22,15 +18,9 @@ public static class AppDataHelper
         //can connect - no database
         //can connect - is database
 
-        if (config.GetValue<bool>("DataInitialization:DropDatabase"))
-        {
-            context.Database.EnsureDeleted();
-        }
+        if (config.GetValue<bool>("DataInitialization:DropDatabase")) context.Database.EnsureDeleted();
 
-        if (config.GetValue<bool>("DataInitialization:MigrateDatabase"))
-        {
-            context.Database.Migrate();
-        }
+        if (config.GetValue<bool>("DataInitialization:MigrateDatabase")) context.Database.Migrate();
 
         if (config.GetValue<bool>("DataInitialization:SeedIdentity"))
         {
@@ -39,8 +29,7 @@ public static class AppDataHelper
 
         if (config.GetValue<bool>("DataInitialization:SeedData"))
         {
-           //TODO
+            //TODO
         }
     }
-
 }
